% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{blankground}
\alias{blankground}
\title{A blank theme, copied mostly or completely from a blog post
that I can't find anymore}
\usage{
blankground()
}
\description{
A blank theme, copied mostly or completely from a blog post
that I can't find anymore
}
