% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MiscellaneousRFunctions.R
\name{cUnique}
\alias{cUnique}
\title{Count the number of unique values in an array}
\usage{
cUnique(x, count.na = FALSE)
}
\arguments{
\item{x}{an array}

\item{count.na}{a boolean, if T, NAs will be counted as unique values}
}
\value{
the number of unique values in x
}
\description{
Returns the number of unique values in an array
}
\examples{
cUnique(x = letters)
> 26
cUnique(x = factor(letters))
> 26
cUnique(x = as.numeric(letters), count.na = F)
> 0
cUnique(x = as.numeric(letters), count.na = T)
> 1
}

