% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{factorConvert}
\alias{factorConvert}
\title{Converts factors into numeric or character arrays}
\usage{
factorConvert(var, to = "numeric")
}
\arguments{
\item{var}{an array of factors to be converted}

\item{to}{a string indicating whether factor should be converted to a character (to = "c") or a numeric (to = "n")}
}
\value{
a vector containing a converted version of var
}
\description{
Converts factors into numeric or character arrays
}
\examples{
> factorConvert(var = as.factor(letters[1:5]), to = "c")
[1] "a" "b" "c" "d" "e"
> factorConvert(var = as.factor(1:5), to = "c")
[1] "1" "2" "3" "4" "5"
> factorConvert(var = as.factor(1:5), to = "n")
[1] 1 2 3 4 5

}
