% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{getKMData}
\alias{getKMData}
\title{Transform survfit object to data.frame}
\usage{
getKMData(survfit)
}
\arguments{
\item{survfit}{is a survfit object}
}
\value{
a data frame containing data from .survfit in a more-easily-plotable format
}
\description{
Transforms a survfit object into a data frame
suitable for generating Kaplan-Meier plots
}
\examples{
> library(survival)
> library(ggplot2)
> survival_fit <- survfit(Surv(time, status) ~ x, data = aml)
> plot_data <- getKMData(survival_fit)
> ggplot(data = plot_data, aes(x = time, y = surv, color = strata)) +
+ geom_line()

}
