% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MiscellaneousRFunctions.R
\name{iqr}
\alias{iqr}
\title{Calculate IQR and IQR width}
\usage{
iqr(x, width = T, na.rm = F)
}
\arguments{
\item{x}{An array of numbers.}
}
\value{
x's IQR or width of x's IQR
}
\description{
Uses default parameters
in quantile() function
}
\examples{
iqr(1:100, width = T)
iqr(1:100, width = F)
}

