% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MiscellaneousRFunctions.R
\name{logit}
\alias{logit}
\title{Calculate the logit of a number}
\usage{
logit(x)
}
\arguments{
\item{x}{A number between 0 and 1.}
}
\value{
log( x / ( 1 - x ) )
}
\description{
Calculate the logit of a number
}
\examples{
logit(0.9)
}

