% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{openFilesInDirectory}
\alias{openFilesInDirectory}
\title{Loads and combines files in a folder that match a specified
string}
\usage{
openFilesInDirectory(
  dir,
  match_string,
  merge = FALSE,
  sep = ",",
  na.strings = "NA",
  header = T,
  fill = T,
  quote = "\\"",
  skip = 0
)
}
\arguments{
\item{dir}{the directory in which to search for files}

\item{match_string}{all files whose names grep match this string will be loaded}

\item{merge}{T/F should loaded files be merged into a data frame}

\item{sep}{field separator character}

\item{na.strings}{a character vector of strings whic are to be interpreted as NA values}

\item{header}{a T/F value indicating whether the file contains the names of the variables as its first line. If missing, the value is determined from the file format: header is set to TRUE if and only if the first row contains one fewer field than the number of columns.}

\item{fill}{T/F If TRUE then in case the rows have unequal length, blank fields are implicitly added}

\item{quote}{Character indicating characters to use for quote argument in read.table}

\item{skip}{the number of lines of the data file to skip before beginning to read data}
}
\value{
a list or data frame containing load data from all files matching match_string in directory
}
\description{
Returns either a list containing all loaded files, or
a data frame containing merged loaded files
}
