% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{poolImpute}
\alias{poolImpute}
\title{Pools estimates from imputed data using Rubin's Rules}
\usage{
poolImpute(estimate, se)
}
\arguments{
\item{estimate}{A set of parameter estimates from multiple
imputed data sets}

\item{se}{A set of standard errors corresponding to the
parameter estimates in estimate}
}
\description{
Pools estimates from imputed data using Rubin's Rules
}
\examples{
poolImpute(1:10, 1:10/5)
>   pooled_coef pooled_se  pooled_p
> 1         5.5  3.409301 0.1403821
}
