% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{printDec}
\alias{printDec}
\title{Print numbers to an arbitrary number of decimal places}
\usage{
printDec(x, dec = 2)
}
\arguments{
\item{x}{A number}

\item{dec}{An integer indicating the number of decimal places to use and round to}
}
\description{
Print numbers to an arbitrary number of decimal places
}
\examples{
printDec(345.23, 4)
[1] "345.2300"

}
