% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{removeCols}
\alias{removeCols}
\title{Remove columns from a data frame}
\usage{
removeCols(.df, .match_string)
}
\arguments{
\item{.df}{is a data frame}

\item{.match_string}{is a string for grepl matching}
}
\value{
.df with columns matched by .match_string removed
}
\description{
Use regexp matching to remove columns from a data frame
}
\examples{
> test <- data.frame(x = c(1,2,3),
                     y = c(4,5,6),
                     xa = c(7,8,9),
                     apple = c(5,5,5))
> removeCols(test, "x")
  y apple
1 4     5
2 5     5
3 6     5
> removeCols(test, "x|a")
  y
1 4
2 5
3 6

}
