% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{removeNAs}
\alias{removeNAs}
\title{Removes rows with NAs in a data frame}
\usage{
removeNAs(df, vars)
}
\arguments{
\item{.df}{a data frame}

\item{.vars}{a charactor vector containing
names of columns to search for NAs}
}
\value{
a data.frame purged of all rows with NAs
 in the columns specified by .vars
}
\description{
Removes rows with NAs
in columns specified by .vars argument
}
\examples{
test_df <- diag(3)
diag(test_df) <- NA
test_df <- data.frame(test_df)
removeNAs(test_df, c("X1", "X3"))
> 1 (33.3\%) missing value(s) found in X1
> 1 (33.3\%) missing value(s) found in X3
> 2 observation(s) with missing values removed from dataset,
> 1 of 3 (33.3\%) remaining
> X1 X2 X3
> 2  0 NA  0
}
