% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortSimulation.R
\name{simulateRandomInterceptModel}
\alias{simulateRandomInterceptModel}
\title{Simulate fixed-slope, random-intercept longitudinal data}
\usage{
simulateRandomInterceptModel(
  n_subjects,
  n_timepoints,
  intercept_mean,
  intercept_sd,
  slope,
  epsilon,
  group = "",
  id_prefix = ""
)
}
\arguments{
\item{n_subjects}{Number of subjects}

\item{n_timepoints}{Number of timepoints}

\item{intercept_mean}{Mean random intercept}

\item{intercept_sd}{Standard deviation of random intercept}

\item{slope}{Slope}

\item{epsilon}{Observation-level error}

\item{group}{Character string defining group name}

\item{id_prefix}{Character to append to ID variable}
}
\description{
Simulate fixed-slope, random-intercept longitudinal data
}
\examples{
n_subjects <- 10
n_timepoints <- 4
intercept_mean <- 0
intercept_sd <- 0.01
slope <- -0.5
epsilon <- 0.01
group <- "testgroup"
id_prefix <- "test"
simulateRandomInterceptModel(n_subjects, n_timepoints, intercept_mean, intercept_sd, slope, epsilon, group, id_prefix)

}
