% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortSimulation.R
\name{simulateSlopeModel}
\alias{simulateSlopeModel}
\title{Simulate fixed-slope, fixed-intercept longitudinal data}
\usage{
simulateSlopeModel(
  n_subjects,
  n_timepoints,
  intercept,
  slope,
  epsilon,
  group = "",
  id_prefix = ""
)
}
\arguments{
\item{n_subjects}{Number of subjects}

\item{n_timepoints}{Number of timepoints}

\item{intercept}{Intercept}

\item{slope}{Slope}

\item{epsilon}{Observation-level error}

\item{group}{Character string defining group name}

\item{id_prefix}{Character to append to ID variable}
}
\description{
Simulate fixed-slope, fixed-intercept longitudinal data
}
\examples{
n_subjects <- 10
n_timepoints <- 4
intercept <- 0
slope <- -0.5
epsilon <- 0.01
group <- "testgroup"
id_prefix <- "test"
simulateSlopeModel(n_subjects, n_timepoints, intercept, slope, epsilon, group, id_prefix)

}
