% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{theme_complete_bw}
\alias{theme_complete_bw}
\title{A completely BW theme, copied mostly or completely from a blog post
that I can't find anymore}
\usage{
theme_complete_bw(base_size = 11, base_family = "")
}
\description{
A completely BW theme, copied mostly or completely from a blog post
that I can't find anymore
}
