% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{theme_facet}
\alias{theme_facet}
\title{A theme for facet plots
#' @description Use regexp matching to remove columns from a data frame}
\usage{
theme_facet()
}
\value{
a theme object
}
\description{
A theme for facet plots
#' @description Use regexp matching to remove columns from a data frame
}
