% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{alogit}
\alias{alogit}
\title{Calculate the inverse logit of a number}
\usage{
alogit(x)
}
\arguments{
\item{x}{A number.}
}
\value{
exp(x) / (1 + exp(x))
}
\description{
Calculate the inverse logit of a number
}
\examples{
alogit(0.9)

}
