% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{iqr}
\alias{iqr}
\title{Calculate IQR and IQR width}
\usage{
iqr(x, width = T, na.rm = F)
}
\arguments{
\item{x}{An array of numbers.}

\item{width}{A boolean indicating whether function should return with or endpoints}

\item{na.rm}{A boolean indicating whether NAs should be removed}
}
\value{
x's IQR or width of x's IQR
}
\description{
Uses default parameters
in quantile() function
}
\examples{
iqr(1:100, width = T)
iqr(1:100, width = F)
}
